fix(cli): version detection, trailing newline, test consolidation
- Read version from package.json via createRequire instead of
  process.env.npm_package_version (which is only set via npm run)
- Add trailing newline after ASCII output for standard Unix behavior
- Consolidate single-test describe blocks into logical groups in
  cli.test.ts and cli-e2e.test.ts
- Extract shared createMockStdout/renderArgs helpers into
  cli-test-helpers.ts to eliminate duplication
